+++ packager/media/base/media_handler_test_base.cc
@@ -73,8 +73,8 @@ class FakeMediaHandler : public MediaHandler {
     stream_data_vector_.push_back(std::move(stream_data));
     return Status::OK;
   }
-  Status OnFlushRequest(int input_stream_index) override { return Status::OK; }
-  bool ValidateOutputStreamIndex(int stream_index) const override {
+  Status OnFlushRequest(size_t input_stream_index) override { return Status::OK; }
+  bool ValidateOutputStreamIndex(size_t stream_index) const override {
     return true;
   }
