+++ packager/media/formats/mp4/mp4_muxer.h
@@ -37,8 +37,9 @@ class MP4Muxer : public Muxer {
   // Muxer implementation overrides.
   Status InitializeMuxer() override;
   Status Finalize() override;
-  Status AddSample(int stream_id, std::shared_ptr<MediaSample> sample) override;
-  Status FinalizeSegment(int stream_id,
+  Status AddSample(size_t stream_id,
+                   std::shared_ptr<MediaSample> sample) override;
+  Status FinalizeSegment(size_t stream_id,
                          std::shared_ptr<SegmentInfo> segment_info) override;
 
   // Generate Audio/Video Track box.
